Q124-001 disable all statistics except "rlimit count"
Disable all statistics except "rlimit count" (which in smt-output mode is
pretty-printed with a dash, i.e. as "rlimit-count") to minimize the amount
of memory needed for caching the Z3 outputs. This commit is meant to be as
non-intrusive as possible; in particular, we not want to disable all
existing statistics individually nor disable new ones when they are added by
the Z3 team.

* src/util/statistics.cpp
(INIT_DISPLAY): explicitly list statistics that GNATprove cares about.

Change-Id: Ia12f68e899d8ed319150d79d59f134030d618139